adds in react-native-paper and reorganizes each page to be under a sub dir
