+++ package.json
@@ -25,11 +25,6 @@
     "@react-navigation/native": "^7.0.14",
     "@react-navigation/stack": "^7.1.1",
     "@reduxjs/toolkit": "^2.5.0",
-    "@tensorflow-models/universal-sentence-encoder": "^1.3.3",
-    "@tensorflow/tfjs": "^4.22.0",
-    "@tensorflow/tfjs-backend-cpu": "^4.22.0",
-    "@tensorflow/tfjs-backend-webgl": "^4.22.0",
-    "@tensorflow/tfjs-core": "^4.22.0",
     "axios": "^1.7.9",
     "eslint": "^8.53.0",
     "glob": "^10.4.5",
@@ -44,7 +39,7 @@
     "react-native-dotenv": "^3.4.11",
     "react-native-gesture-handler": "^2.21.2",
     "react-native-modal": "^13.0.1",
-    "react-native-paper": "^5.12.5",
+    "react-native-paper": "^5.13.1",
     "react-native-reanimated": "^3.16.6",
     "react-native-safe-area-context": "^5.0.0",
     "react-native-screens": "^4.4.0",
@@ -72,6 +67,7 @@
     "@types/node": "^22.10.5",
     "@types/react": "^18.3.18",
     "@types/react-native": "^0.72.8",
+    "@types/react-native-vector-icons": "^6.4.18",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.6.3",
     "jest": "^29.6.3",
