+++ src/Pages/HomePage.tsx
@@ -1,0 +1,11 @@
+import { Text,View } from "react-native";
+
+function HomePage() {
+    return ( 
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+        <Text>Home!</Text>
+    </View>
+       );
+}
+
+export default HomePage;
